binance_bot/requirements.txt
ccxt
pandas
pandas_ta
python-dotenv
asyncio
loguru
numpy
rich
